2. strncpy

+++ 0x06-pointers_arrays_strings/2-strncpy.c
@@ -7,9 +7,9 @@
  * @n: max number of characters to be copied
  * Return: pointer to the destination string
  */
-char *_strncpy(char *dest, char *src, size_t n)
+char *_strncpy(char *dest, char *src, int n)
 {
-	size_t i;
+	int i;
 
 	for (i = 0; i < n && src[i] != '\0'; i++)
 	{
